requirements.txt
pyyaml
prettytable
pylint
jmespath
jc
pytest-xdist
pytest-cov
pytest
